% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_vars}
\alias{lookup_vars}
\title{Variables to rename}
\format{
A named \verb{character vector}
}
\usage{
lookup_vars
}
\description{
Variables to recode with old and new names
}
\keyword{datasets}
