% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_club_data.R
\name{historical_club_data}
\alias{historical_club_data}
\title{name}
\usage{
historical_club_data(leagues = NULL, seasons = NULL)
}
\arguments{
\item{x}{..}
}
\value{
...
}
\description{
This function...
}
\author{
Jeppe Vierø
}
